add patientlist to nurse folder, add routing to pateintlist
