+++ frontend/bingungcore/src/App.js
@@ -9,6 +9,7 @@ import AppointmentSchedule from './pages/Admin/AppointmentSchedule';
 import PatientManagement from './pages/Admin/PatientManagement';
 import RoomManagement from './pages/Admin/RoomManagement';
 import PatientCheckup from './pages/Doctor/PatientCheckup';
+import PatientList from './pages/Nurse/PatientList';
 
 
 // Main App Component
@@ -24,6 +25,7 @@ export default function App() {
                 <Route path="/admin/appointmentSchedule" element={<AppointmentSchedule/>}></Route>
                 <Route path="/admin/patientManagement" element={<PatientManagement/>}></Route>
                 <Route path="/admin/roomManagement" element={<RoomManagement/>}></Route>
+                <Route path="/nurse/patientList" element={<PatientList />} />
                 <Route path="/doctor/patientCheckup/:id" element={<PatientCheckup />} />
 
             </Routes>
